% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omueconApp-package.R
\docType{package}
\name{omueconApp-package}
\alias{omueconApp}
\alias{omueconApp-package}
\title{omueconApp: Small Shiny App for Moodle Pages}
\description{
Small Shiny App That Helps You Convert a Markdown Document into an HTML Fragment that's Ready for Copy-and-Pasting into Moodle.
}
\author{
\strong{Maintainer}: Kenji Sato \email{mail@kenjisato.jp}

}
\keyword{internal}
